CLN: Indirect all smoothing/differentiation/udot_integration
Different libraries meant different
things happened:
- self.differentiation_method
- FiniteDifference
- convert_udot_integral

These were all about calculating the left hand side of the SINDy objective.
That means that we have a common abstraction, and we can use that to remove a
lot of conditional behavior by assigning calculate_trajectory to the correct
method, once.

Add that assignment to BaseFeatureLibrary, PDELibrary, WeakPDELibrary, and
GeneralizedLibrary.  Trade off: GeneralizedLibrary now needs to import
PDELibrary, which seems an acceptable deepening of internal package dependency.

Because the feature libraries also deal with different shapes of inputs, we
can do the same with validate_input().  Maybe ideally there's a better
abstraction of "well formatted data" that works for all SINDy libraries, but in
the interim, handle this function within the class and then at runtime,
dispatch to the appropriate class method.