CLN: Move AxesArray to utils
If this function is in utils, then it would be almost circular to import
it from feature library; it would work because axes.py has no internal
dependencies, but then a module in utils would import from a module in
feature_library and vice versa.

It may be a bit awkward that it carries with it DefaultShapedInputsMixin and
PDEShapedInputsMixin.