ENH: Add SampleConcatter step to SINDy Pipeline.
Also require ability to unshape the data so predictions are same shape as inputs.
Pipeline does not allow this (see [here](https://stackoverflow.com/questions/25539311))
so need to add feature_library.reshape_samples_to_spatial_grid().

Could be a function rather than a method, not sure how a GeneralizedLibrary would work
which doesn't inherit spatial grid.  Solution is probably to calculate the spatial grid
outside the library and pass as an argument.