CLN: Remove fit_intercept in optimizers
Fixes #338

Initialize fit_intercept in super() as False

Prevent the _tragic_ :sob: case of fitting both an intercept and a constant
term, which could only otherwise be achieved with higher coupling and more
tests.

Users can specify a constant term in the feature libraries using include_bias,
where it exists, or by adding a constant term to a custom library and
concatenating.

This change incurs a minor performance hit that disappears with scale, based
upon how sklearn.LinearRegression fit the coefficient separately from the
regression matrices.