CLN: Corrected "failed to converge" SR3 warning
The "failed to converge warning" was emitted whenever max iteration was reached,
regardless of whether convergence was achieved.  The previous test only was able
to assert convergence warning because it was written to successfully converge
after 1 iteration, with a max iteration of 1.

Also, removed test_sr3_warn and incorporated into test_fit_warn, with which it
was 95% redundant already.  This also required setting some variables outside
a loop in order to access them later - a required practice for mypy and other
type checkers.