ENH(sr3): refactor threshold and thresholder property (#548)
As originally written, "threshold" parameter in SR3 classes referred to the l0 norm. 
Appendix A of the SR3 paper included a conversion from l0 threshold to l0 weight,
which was then used in prox, etc.  However, the name threshold stayed around, even
as it was used as a regularizer.  It even applied the l0-specific conversion to l1 and l2
thresholds.

This PR aims to generalize SR3 to other norm by moving threshold-to-weight
calculation to a separate function that the user can call on: 
`calculate_l0_weight()`