BLD: Bump/remove extra dependencies
This bumps the cvxpy version to the first with the default solver Clarabel,
in place of ECOS.  This originally caused some issues due to the different
kwargs expected by the default solver, but upstream packages seem to have
fixed that.

The change also removes restrictions on jax/scipy in the sbr extra.  Scipy
restriciton only existed for Arviz compatability.  Arviz is only used in a
notebook, so that dependency will be part of the notebook's requirements.
However, we import directly from scipy, so it should be a named dependency.