DOC: Make examples directive create a section at correct level.
Previously, the directive worked, but created a heading one level lower than
the previous section.  What we want, however, is a series of these directives
not to nest, but to be siblings of each other under the level of the enclosing
section.

When the directive comes after a heading of the appropriate level, this works.
However, when it comes directly after the title level, we *do* want a lower
section level.