DOC: disambiguate examples from different branches of same repo
Previously, only the repo and file names were used to assign files to examples,
as initiated by the custom directive.  When multiple examples use different
branches of the same repo, each would overwrite filenames of previous examples.
Now, example ref (commit) is used in the filename to prevent this.

Also added ref to the "original source" links for each example.