BLD: Replace sphinx-apidoc with sphinx-autosummary
The goal of this commit is to allow building a bit more sensible docs.
Instead of a massive file with modules and submodules and classes all
documented on the same page, this has one top-levelobject
(module, class, or function) documented on each web page, akin to scikitlearn
e.g. https://scikit-learn.org/stable/api/sklearn.multiclass.html

There is one significant practical decision to make: we need to be clear about
from where we expect an object to be imported.  E.g. I've chosen to hide
documentation of modules like sr3.py and instead document all optimizer classes
under pysindy.optimizers.  We can't do it in two places.  This also means
that pysindy/pysindy.py became pysindy/_core.py, so that SINDy is documented
in the main pysindy package namespace.  There's some other small flattening,
but largely, I expect to hide more and more modules with an underscore as time
goes on, so that helper functions don't end up as part of a backwards
compatability contract.

This isn't the final word in the page layouts.  There's a lot more visual
improvements that can be made.  But this commit should solve the technical
problems and allow us to tweak from here.